feat: refacto title-etablissement
