+++ components/etablissement-section/index.tsx
@@ -5,7 +5,6 @@ import BreakPageForPrint from '#components-ui/print-break-page';
 import { PrintNever } from '#components-ui/print-visibility';
 import { Tag } from '#components-ui/tag';
 import AvisSituationLink from '#components/avis-situation-link';
-import { EtablissementDescription } from '#components/etablissement-description';
 import { Section } from '#components/section';
 import { CopyPaste, TwoColumnTable } from '#components/table/simple';
 import TVACell from '#components/tva-cell';
@@ -19,6 +18,10 @@ import {
   getNomComplet,
 } from '#models/statut-diffusion';
 import { formatDate, formatSiret } from '#utils/helpers';
+import {
+  getCompanyLabel,
+  getCompanyPronoun,
+} from '#utils/helpers/get-company-page-title';
 import { ISession } from '#utils/session';
 
 type IProps = {
@@ -36,11 +39,14 @@ const EtablissementSection: React.FC<IProps> = ({
   withDenomination,
   session,
 }) => {
+  const companyType = `${getCompanyPronoun(
+    uniteLegale
+  ).toLowerCase()}${getCompanyLabel(uniteLegale)}`;
   const data = [
     ...(withDenomination
       ? [
           [
-            'Dénomination de l’unité légale',
+            `Dénomination de ${companyType}`,
             getNomComplet(uniteLegale, session),
           ],
           [
@@ -56,7 +62,7 @@ const EtablissementSection: React.FC<IProps> = ({
               )}
               {' ( '}
               <a key="entite" href={`/entreprise/${uniteLegale.chemin}`}>
-                → voir la page de l’unité légale
+                → voir la page de {companyType}
               </a>
               {' )'}
             </>,
@@ -113,7 +119,7 @@ const EtablissementSection: React.FC<IProps> = ({
         ]
       : []),
     [
-      'Activité principale de l’unité légale (NAF/APE)',
+      `Activité principale de ${companyType} (NAF/APE)`,
       uniteLegale.libelleActivitePrincipale,
     ],
     [
@@ -138,13 +144,6 @@ const EtablissementSection: React.FC<IProps> = ({
 
   return (
     <>
-      {!usedInEntreprisePage && (
-        <EtablissementDescription
-          etablissement={etablissement}
-          uniteLegale={uniteLegale}
-          session={session}
-        />
-      )}
       <Section
         title={
           usedInEntreprisePage
